小项目/unet-buiding/requirements.txt
numpy==1.22.0
torch==1.13.1
torchvision==0.11.1
Pillow
